Consolidate account authentication into UserAccountController
