+++ src/READUS/Controllers/UserAccountController.cs
@@ -1,11 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
+using System.Security.Claims;
+using System.Text;
 using System.Threading.Tasks;
 using DomainObjects;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
 using READUS.Commands;
+using READUS.Models;
 using Storage;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
@@ -24,7 +30,7 @@ namespace READUS.Controllers
         }
 
         // POST api/<controller>
-        [HttpPost]
+        [HttpPost("account")]
         [AllowAnonymous]
         public IActionResult CreateAccount([FromBody]CreateAccountCommand account)
         {
@@ -36,6 +42,51 @@ namespace READUS.Controllers
             this.dataContext.Users.Add(newUser);
 
             return Ok(newUser.Id);
+        }
+
+        [HttpPost("login")]
+        [AllowAnonymous]
+        public IActionResult Login([FromBody]AuthRequest authRequest)
+        {
+            if (authRequest == null)
+                return BadRequest();
+
+            var encryptedPassword = authRequest.Password; // var encryptedPassword = SomeHash(authRequest.Password + salt)
+
+            var user = this.dataContext.Users
+                .GetWhere(x => x.Username == authRequest.Username && x.Password == encryptedPassword)
+                .FirstOrDefault();
+
+            if (user == null)
+                return BadRequest("Username or password is incorrect");
+
+            var token = GenerateJwt(user);
+
+            var cookieOptions = new CookieOptions
+            {
+                Domain = this.Request.Host.Host,
+                Expires = DateTimeOffset.Now.AddMinutes(5),
+                HttpOnly = true,
+                Secure = false // will be true for production usage
+            };
+
+            this.Response.Cookies.Append("Authorization", token, cookieOptions);
+
+            return Ok();
+        }
+
+        private string GenerateJwt(User user)
+        {
+            var jwtHandler = new JwtSecurityTokenHandler();
+            var key = Encoding.ASCII.GetBytes("this is the key that I will use for development");
+            var tokenDescriptor = new SecurityTokenDescriptor
+            {
+                Subject = new ClaimsIdentity(new[] { new Claim("id", user.Id.ToString()) }),
+                Expires = DateTime.UtcNow.AddMinutes(5),
+                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
+            };
+            var token = jwtHandler.CreateToken(tokenDescriptor);
+            return jwtHandler.WriteToken(token);
         }
     }
 }
